Move jsonrpc2 into lsp
