+++ internal/lsp/jsonrpc2/stream.go
