+++ internal/lsp/protocol/log.go
@@ -9,7 +9,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/nokia/ntt/internal/jsonrpc2"
+	"github.com/nokia/ntt/internal/lsp/jsonrpc2"
 )
 
 type loggingStream struct {
