+++ internal/lsp/protocol/tsclient.go
@@ -11,7 +11,7 @@ import (
 	"context"
 	"encoding/json"
 
-	"github.com/nokia/ntt/internal/jsonrpc2"
+	"github.com/nokia/ntt/internal/lsp/jsonrpc2"
 	"github.com/nokia/ntt/internal/telemetry/log"
 	"github.com/nokia/ntt/internal/xcontext"
 )
