+++ internal/lsp/telemetry.go
@@ -9,7 +9,7 @@ import (
 	"encoding/json"
 	"time"
 
-	"github.com/nokia/ntt/internal/jsonrpc2"
+	"github.com/nokia/ntt/internal/lsp/jsonrpc2"
 	"github.com/nokia/ntt/internal/lsp/telemetry"
 	"github.com/nokia/ntt/internal/telemetry/trace"
 )
